added txpetscgpu stuff for testing, compiles not yet tested
Hg-commit: b056329a6646524a5192ed7d8dc41c52402951cb

+++ src/vec/vec/impls/seq/seqcusp/cuspvecimpl.h
@@ -46,7 +46,7 @@ extern PetscErrorCode VecCUSPAllocateCheck_Public(Vec);
 extern PetscErrorCode VecCUSPCopyToGPUSome_Public(Vec,CUSPINTARRAYCPU*,CUSPINTARRAYGPU*);
 
 extern PetscBool  synchronizeCUSP;
-#define CHKERRCUSP(err) if (err != CUBLAS_STATUS_SUCCESS) SETERRQ1(PETSC_COMM_SELF,PETSC_ERR_LIB,"CUSP error %d",err)
+#define CHKERRCUSP(err) if (((int)err) != (int)CUBLAS_STATUS_SUCCESS) SETERRQ1(PETSC_COMM_SELF,PETSC_ERR_LIB,"CUSP error %d",err)
 
 #define VecCUSPCastToRawPtr(x) thrust::raw_pointer_cast(&(x)[0])
